requirements.txt
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.4.2
Flask==3.0.0
Flask-Cors==4.0.0
google-api-core==2.15.0
google-auth==2.26.1
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pymongo==4.6.1
requests==2.31.0
rsa==4.9
urllib3==2.1.0
Werkzeug==3.0.1
